emul: Initial commit of graphics API
The graphics API is used for rendering pixels and text to a simple
OS window, when the underlying emulated code calls to do so. For
example, when trying to show some text, the program will write some
bytes to memory which would be marked with a PAGE_TEXT flag, calling
the assigned handler to then push the text to the render pipeline on
a seperate thread.